refactor(phd): phd frontend routing
