+++ client/src/views/Phd/index.tsx
@@ -1,0 +1,9 @@
+import { Navigate } from "react-router-dom";
+
+const Phd = () => {
+  return (
+    <Navigate to="/phd/notional-supervisor" />
+  );
+};
+
+export default Phd;
